refactor: devide directory of alarm screen and early late screen
