+++ lib/presentation/early_late/components/early_late_message_image_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:on_time_front/presentation/alarm/bloc/ealy_late_screen/early_late_screen_bloc.dart';
+import 'package:on_time_front/presentation/early_late/bloc/early_late_screen_bloc.dart';
 
 class EarlyLateMessageImageWidget extends StatelessWidget {
   final double screenHeight;
