+++ lib/presentation/early_late/screens/early_late_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:on_time_front/presentation/alarm/bloc/ealy_late_screen/early_late_screen_bloc.dart';
-import 'package:on_time_front/presentation/alarm/components/early_late_screen/check_list_box_widget.dart';
-import 'package:on_time_front/presentation/alarm/components/early_late_screen/early_late_message_image_widget.dart';
+import 'package:on_time_front/presentation/early_late/bloc/early_late_screen_bloc.dart';
+import 'package:on_time_front/presentation/early_late/components/check_list_box_widget.dart';
+import 'package:on_time_front/presentation/early_late/components/early_late_message_image_widget.dart';
 import 'package:on_time_front/presentation/shared/components/button.dart';
 import 'package:on_time_front/presentation/shared/utils/time_format.dart';
